+++ post_office/tests/test_mail.py
@@ -1,4 +1,5 @@
 from datetime import date, datetime, timedelta
+from unittest.mock import patch
 
 from django.core import mail
 from django.core.files.base import ContentFile
@@ -428,3 +429,14 @@ class MailTest(TestCase):
         email.refresh_from_db()
         self.assertEqual(email.status, STATUS.failed)
         settings.POST_OFFICE = previous_settings
+
+    @patch('post_office.signals.email_queued.send')
+    def test_backend_signal(self, mock):
+        """
+        Check that the post_office signal handler is fired
+        """
+        email = send(recipients=['a@example.com'],
+                     sender='from@example.com', message='message',
+                     subject='subject')
+        mock.assert_called_once_with(sender=Email, emails=[email])
+